requirements.txt
git+https://github.com/Rapptz/discord.py.git
git+https://github.com/Gorialis/jishaku@master
python-dotenv~=0.15.0
pytimeparse==1.1.8
pymongo
asyncio
dnspython
aiohttp
kitsu-wrapper
timeago
psutil
youtube_dl
pynacl
animec
